requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.16
django-shopping-cart==0.1
idna==3.4
Pillow==9.4.0
pytz==2022.7
razorpay==1.3.0
requests==2.28.2
sqlparse==0.4.3
typing-extensions==4.4.0
urllib3==1.26.14
